some upd

+++ Sources/ServiceLayer/ProfileService/Models/Bonus.swift
@@ -11,7 +11,7 @@ public struct Bonus: Codable, JSONParsable {
     public let currency: String
     public let amount: Double
     public let baseCurrency: String
-    public let amountInBaseCurrency: Int
+    public let amountInBaseCurrency: Double
     
     enum CodingKeys: String, CodingKey {
         case currency, baseCurrency, name
